requirements.txt
aws-cdk-lib==2.101.1
constructs>=10.0.0,<11.0.0

boto3~=1.28.65
urllib3~=2.0.7